chore: Ignore .markdownlintrc
Add .markdownlintrc to the .gitignore file to prevent it from being committed to the repository. This file contains configuration settings for markdownlint, which is used to enforce style guidelines for Markdown files. It's not necessary to commit this file as it's specific to the local environment and can be easily recreated.